+++ pkg/integrations/dnsmasq_exporter/dnsmasq_exporter.go
@@ -11,6 +11,7 @@ import (
 
 // DefaultConfig is the default config for dnsmasq_exporter.
 var DefaultConfig Config = Config{
+	Common:         config.DefaultCommon,
 	DnsmasqAddress: "localhost:53",
 	LeasesPath:     "/var/lib/misc/dnsmasq.leases",
 }