+++ pkg/integrations/memcached_exporter/memcached_exporter.go
@@ -12,6 +12,7 @@ import (
 
 // DefaultConfig is the default config for memcached_exporter.
 var DefaultConfig Config = Config{
+	Common:           config.DefaultCommon,
 	MemcachedAddress: "localhost:11211",
 	Timeout:          time.Second,
 }